moved middleware to its own modules
